pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
selenium = "^4.21.0"
webdriver-manager = "^4.0.1"
undetected-chromedriver = "^3.5.5"
beautifulsoup4 = "^4.12.3"
pandas = "^2.2.2"
sqlalchemy = "^2.0.30"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.1"
psycopg2-binary = "^2.9.9"
requests = "^2.32.3"
replit = "^4.0.0"
crewai = "^0.30.11"
fpdf = "^1.7.2"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"